more lint fixes
