+++ build-system/compile/compile.js
@@ -28,11 +28,11 @@ const {
   handleCompilerError,
   handleTypeCheckError,
 } = require('./closure-compile');
+const {CLOSURE_SRC_GLOBS} = require('../sources');
 const {isTravisBuild} = require('../travis');
 const {shortenLicense, shouldShortenLicense} = require('./shorten-license');
 const {singlePassCompile} = require('./single-pass');
 const {VERSION: internalRuntimeVersion} = require('../internal-version');
-const {CLOSURE_SRC_GLOBS} = require('../tasks/helpers');
 
 const isProdBuild = !!argv.type;
 const queue = [];
